backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
pywebpush==2.0.3
